+++ Sources/BraintreeShopperInsights/BTShopperInsightsAnalytics.swift
@@ -5,10 +5,10 @@ enum BTShopperInsightsAnalytics {
     // MARK: - Merchant Triggered Events
       
     static let payPalPresented = "shopper-insights:paypal-presented"
-    static let payPalSelected = "shopper-insights:paypal-selected"
     static let venmoPresented = "shopper-insights:venmo-presented"
-    static let venmoSelected = "shopper-insights:venmo-selected"
-    
+
+    static let buttonSelected = "shopper-insights:button-selected"
+
     // MARK: - SDK Triggered Events
     
     static let recommendedPaymentsStarted = "shopper-insights:get-recommended-payments:started"